Updating services page
