+++ src/app/page.tsx
@@ -1,9 +1,10 @@
 import Image from "next/image";
 import NavLink from "@/src/ui/nav-links";
-import Footer from "@/src/ui/copyright-footer";
+import Footer from "@/src//ui/copyright-footer";
 import SmallContactBar from "@/src/ui/small-contact-bar";
 import ImageCarousel from "@/src/ui/image-carousel";
 import Card from "@/src/ui/card";
+import React from "react";
 
 export default function Home() {
   return (
